Try to simplify the lexer implementation
Emphasis on "try".

Came back to the old code after some months'
break and found a bug with the pushback char
implementation, as well as getting confused
about some of the implementation and breaking
it.

Now there's less logic in the states themselves,
especially with iterating the token character
or setting the pushback character.

I, uh... am still not satisfied, but it's...
something.